lang_chain_agnets/requirements.txt
flask
python-dotenv
langchain
langgraph
openai
duckduckgo-search
gunicorn
langchain-openai
langchain-community
duckduckgo-search